Refactor: Outsource contributor person event handler to its own module
- Moved contributor person logic into a dedicated file (contributor-person.js)
- Added full JSDoc documentation for clarity and maintainability
- Refactored role and affiliation input fields for improved structure and readability.
