+++ js/buttons.js
@@ -301,98 +301,6 @@ $(document).ready(function () {
     setupContactPersonToggle();
   });
 
-  /**
-   * Event handler for adding a new contributor person row.
-   * Clones the first contributor person row, resets all input fields,
-   * updates IDs to ensure uniqueness, and initializes all necessary components.
-   * 
-   * @event #button-contributor-addperson#click
-   * @requires jQuery
-   * @requires Tagify
-   */
-  $("#button-contributor-addperson").click(function () {
-    const contributorGroup = $("#group-contributorperson");
-    const firstContributorRow = contributorGroup.children().first();
-    const uniqueSuffix = new Date().getTime();
-
-    const newContributorRow = firstContributorRow.clone();
-
-    newContributorRow.find("input").val("").removeClass("is-invalid is-valid");
-    newContributorRow.find(".tagify").remove();
-    newContributorRow.find(".invalid-feedback, .valid-feedback").css("display", "");
-    replaceHelpButtonInClonedRows(newContributorRow);
-    newContributorRow.find(".row-label").hide();
-    newContributorRow.find("input").removeAttr("required");
-
-    // Completely replace the role input field container with fresh HTML
-    const roleFieldContainer = newContributorRow.find("#input-contributor-personrole").closest(".input-group");
-    const roleFieldHtml = `
-    <div class="input-group has-validation">
-      <input name="cbPersonRoles[]" id="input-contributor-personrole${uniqueSuffix}"
-        class="form-control tagify--custom-dropdown input-with-help input-right-no-round-corners"
-        data-translate-placeholder="general.roleLabel" />
-      <span class="input-group-text"><i class="bi bi-question-circle-fill"
-        data-help-section-id="help-contributor-role"></i></span>
-      <div class="invalid-feedback" data-translate="general.PleaseChoose">Please choose</div>
-    </div>
-  `;
-    roleFieldContainer.replaceWith(roleFieldHtml);
-
-    // Also replace the affiliation input field with fresh HTML
-    const affFieldContainer = newContributorRow.find("#input-contributorpersons-affiliation").closest(".input-group");
-    const affFieldHtml = `
-    <div class="input-group has-validation">
-      <input type="text" name="cbPersonAffiliations[]" id="input-contributorpersons-affiliation${uniqueSuffix}"
-        class="form-control input-with-help input-right-no-round-corners" 
-        data-translate-placeholder="general.affiliation" />
-      <input type="hidden" name="cbPersonRorIds[]" id="input-contributor-personrorid${uniqueSuffix}" />
-      <span class="input-group-text"><i class="bi bi-question-circle-fill"
-        data-help-section-id="help-contributor-affiliation"></i></span>
-      <div class="invalid-feedback" data-translate="general.PleaseChoose">Please choose</div>
-    </div>
-  `;
-    affFieldContainer.replaceWith(affFieldHtml);
-
-    // Update remaining input IDs
-    newContributorRow
-      .find("#input-contributor-orcid")
-      .attr("id", "input-contributor-orcid" + uniqueSuffix);
-    newContributorRow
-      .find("#input-contributor-lastname")
-      .attr("id", "input-contributor-lastname" + uniqueSuffix);
-    newContributorRow
-      .find("#input-contributor-firstname")
-      .attr("id", "input-contributor-firstname" + uniqueSuffix);
-
-    // Update label references
-    const labelMappings = ["orcid", "lastname", "firstname"];
-    labelMappings.forEach(label => {
-      newContributorRow
-        .find(`label[for='input-contributor-${label}']`)
-        .attr("for", `input-contributor-${label}${uniqueSuffix}`);
-    });
-
-    newContributorRow.find(".addContributorPerson").replaceWith(removeButton);
-    contributorGroup.append(newContributorRow);
-
-    // Initialize Tagify for roles
-    setupRolesDropdown(["person", "both"], `#input-contributor-personrole${uniqueSuffix}`);
-
-    // Check if affiliationsData is available in global scope
-    if (window.affiliationsData && Array.isArray(window.affiliationsData)) {
-      autocompleteAffiliations(
-        `input-contributorpersons-affiliation${uniqueSuffix}`,
-        `input-contributor-personrorid${uniqueSuffix}`,
-        window.affiliationsData
-      );
-    }
-
-    newContributorRow.on("click", ".removeButton", function () {
-      $(this).closest(".row").remove();
-      checkMandatoryFields();
-    });
-  });
-
 
   /**
   * Global variable to keep track of unique tsc-row-ids.
